+++ packages/uniforms/src/bridges/SimpleSchema2Bridge.js
@@ -91,7 +91,20 @@ export default class SimpleSchema2Bridge extends Bridge {
 
         invariant(definition, 'Field not found in schema: "%s"', name);
 
-        return {...definition, ...definition.type[0]};
+        const merged = {
+            ...definition,
+            ...definition.type[0]
+        };
+
+        // aldeed/node-simple-schema#27
+        if (merged.autoValue && (
+            merged.autoValue.name === 'defaultAutoValueFunction' ||
+            merged.autoValue.toString().indexOf('$setOnInsert:') !== -1 // FIXME: Hack.
+        )) {
+            merged.defaultValue = merged.autoValue.call({operator: null});
+        }
+
+        return merged;
     }
 
     getInitialValue (name, props = {}) {